오브젝트 충돌설정 1차

+++ WinAPI/ContentsProject/Monster.h
@@ -304,7 +304,7 @@ protected:
 	APlayer* Player = nullptr;
 
 	// Random
-	int RandomSeed = 0;
+	__int64 RandomSeed = 0;
 	UEngineRandom MonsterRandom;
 private:
